+++ class/CheckMailCellDup.php
@@ -1,0 +1,53 @@
+<?php
+    class CheckMailCellDup {
+        public $checkMail;
+        public $checkCell;
+        public $checkSafe;
+
+
+
+        // getter checkMail
+        public function getCheckMail()
+        {
+            return $this->checkMail;
+        }
+
+        // setter checkMail
+        public function setCheckMail($checkMail)
+        {
+            $this->checkMail = $checkMail;
+
+            return $this;
+        }
+
+        
+        // getter checkCell
+        public function getCheckCell()
+        {
+            return $this->checkCell;
+        }
+
+        // setter checkCell
+        public function setCheckCell($checkCell)
+        {
+            $this->checkCell = $checkCell;
+
+            return $this;
+        }
+
+        // getter checkSafe
+        public function getCheckSafe()
+        {
+            return $this->checkSafe;
+        }
+
+        // Setter checkSafe
+        public function setCheckSafe($checkSafe)
+        {
+            $this->checkSafe = $checkSafe;
+
+            return $this->checkSafe;
+        }
+
+        
+    }