+++ class/classes.php
@@ -1,0 +1,11 @@
+<?php
+    require("class/Gallery.php");
+
+    require("class/Service.php");
+
+    // CheckMailCellDup.php = Classe usada para o objeto de verificação de dados duplicado
+    // em um cadastro de um novo usuário
+    require("class/CheckMailCellDup.php");
+
+    // User.php = classe usada para objeto de usuário
+    require("class/User.php");